+++ VS.Squirrel/SquirrelPackagerPackage.cs
@@ -17,6 +17,7 @@ using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.Win32;
 using EnvDTE80;
 using EnvDTE;
+using AutoSquirrel.Services.Helpers;
 
 namespace AutoSquirrel
 {
@@ -41,7 +42,6 @@ namespace AutoSquirrel
     [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)] // Info on this package for Help/About
     [ProvideMenuResource("Menus.ctmenu", 1)]
     [ProvideToolWindow(typeof(SquirrelPackager))]
-    [ProvideToolWindowVisibility(typeof(SquirrelPackager), VSConstants.UICONTEXT.SolutionExistsAndFullyLoaded_string)]
     [Guid(SquirrelPackagerPackage.PackageGuidString)]
     [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly", Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
@@ -75,6 +75,7 @@ namespace AutoSquirrel
             _dte = GetService(typeof(DTE)) as DTE2;
             SquirrelPackagerCommand.Initialize(this);
             base.Initialize();
+            VSHelper.ProjectIsValid.Value = false;
         }
     }
 }